include new attributes, and add methods for calculating totals directly in the DTO

+++ Backend/src/main/java/com/gabriel/Backend/dto/ShoppingCartDto.java
@@ -7,6 +7,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.util.Set;
 
 @Getter
@@ -17,10 +18,8 @@ public class ShoppingCartDto {
     private Long id;
 
     private CustomerDto customer;
-    private double totalPrice;
-
+    private BigDecimal totalPrice;
     private int totalItems;
-
     private Set<CartItemDto> cartItems;
 
 }
